fix doc

+++ screwbox-core/src/main/java/dev/screwbox/core/window/internal/CursorLockInSupport.java
@@ -11,11 +11,11 @@ import static java.lang.Math.clamp;
 public record CursorLockInSupport(Robot robot, Mouse mouse) {
 
     public void lockInCursor(final ScreenBounds bounds, final int padding) {
-        final int minX = bounds.offset().x() + padding;
-        final int maxX = bounds.offset().x() + bounds.width() - padding;
+        final long minX = bounds.offset().x() + padding;
+        final long maxX = bounds.offset().x() + bounds.width() - padding;
 
-        final int minY = bounds.offset().y() + padding;
-        final int maxY = bounds.offset().y() + bounds.height() - padding;
+        final long minY = bounds.offset().y() + padding;
+        final long maxY = bounds.offset().y() + bounds.height() - padding;
 
         final Offset targetMouse = Offset.at(
                 clamp(mouse.offset().x() + bounds.offset().x(), minX, maxX),
